PPA-NSP
Upload of the NSP-PPA Algorithm that work with the NSP Library set of problems.

The implementation solve the hard constraints of minimum required allocation per day-shift, also support of optimization of personell schedulling preferences and number of  skilled personell per shift-day.
The result return the entire schedule of all the personel, to which shift-day have been assigned.